(FMT): move createBlock to fmt model builder
