+++ src/main/kotlin/org/rust/ide/formatter/RustFormattingModelBuilder.kt
@@ -1,23 +1,35 @@
 package org.rust.ide.formatter
 
-import com.intellij.formatting.FormattingModel
-import com.intellij.formatting.FormattingModelBuilder
-import com.intellij.formatting.FormattingModelProvider
-import com.intellij.formatting.Indent
+import com.intellij.formatting.*
 import com.intellij.lang.ASTNode
 import com.intellij.openapi.util.TextRange
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiFile
 import com.intellij.psi.codeStyle.CodeStyleSettings
 import org.rust.ide.formatter.blocks.RustFmtBlock
+import org.rust.ide.formatter.blocks.RustMacroArgFmtBlock
+import org.rust.lang.core.psi.RustCompositeElementTypes.MACRO_ARG
 
 class RustFormattingModelBuilder : FormattingModelBuilder {
     override fun getRangeAffectingIndent(file: PsiFile?, offset: Int, elementAtOffset: ASTNode?): TextRange? = null
 
     override fun createModel(element: PsiElement, settings: CodeStyleSettings): FormattingModel {
         val ctx = RustFmtContext.create(settings)
-        val block = RustFmtBlock.createBlock(element.node, null, Indent.getNoneIndent(), null, ctx)
+        val block = createBlock(element.node, null, Indent.getNoneIndent(), null, ctx)
         /** / com.intellij.formatting.FormattingModelDumper.dumpFormattingModel(block, 2, System.err) // */
         return FormattingModelProvider.createFormattingModelForPsiFile(element.containingFile, block, settings)
+    }
+
+    companion object {
+        fun createBlock(
+            node: ASTNode,
+            alignment: Alignment?,
+            indent: Indent?,
+            wrap: Wrap?,
+            ctx: RustFmtContext
+        ): ASTBlock = when (node.elementType) {
+            MACRO_ARG -> RustMacroArgFmtBlock(node, alignment, indent, wrap, ctx)
+            else -> RustFmtBlock(node, alignment, indent, wrap, ctx)
+        }
     }
 }
